move common code to NotebookLockingUtils
